# src/main/java/ru/miphi/otpauthservicehw/dto/response/LoginResponse.java
package ru.miphi.otpauthservicehw.dto.response;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Builder;

@Builder
public record LoginResponse(

        @JsonProperty(value = "token", required = true)
        String token

) {}